+++ dffml/util/testing/source.py
@@ -5,7 +5,7 @@ import abc
 import random
 import tempfile
 
-from ...repo import Repo, RepoPrediction
+from ...record import Record, RecordPrediction
 from ..asynctestcase import AsyncTestCase
 
 
@@ -20,7 +20,7 @@ class SourceTest(abc.ABC):
     >>> from dffml.util.asynctestcase import AsyncTestCase
     >>> class TestCustomSQliteSource(SourceTest, AsyncTestCase):
     >>>     async def setUpSource(self):
-    >>>         return MemorySource(MemorySourceConfig(repos=[Repo('a')]))
+    >>>         return MemorySource(MemorySourceConfig(records=[Record('a')]))
     """
 
     @abc.abstractmethod
@@ -30,7 +30,7 @@ class SourceTest(abc.ABC):
     async def test_update(self):
         full_key = "0"
         empty_key = "1"
-        full_repo = Repo(
+        full_record = Record(
             full_key,
             data={
                 "features": {
@@ -40,13 +40,13 @@ class SourceTest(abc.ABC):
                     "SepalWidth": 2.7,
                 },
                 "prediction": {
-                    "target_name": RepoPrediction(
+                    "target_name": RecordPrediction(
                         value="feedface", confidence=0.42
                     )
                 },
             },
         )
-        empty_repo = Repo(
+        empty_record = Record(
             empty_key,
             data={
                 "features": {
@@ -62,41 +62,43 @@ class SourceTest(abc.ABC):
         async with source as testSource:
             # Open, update, and close
             async with testSource() as sourceContext:
-                await sourceContext.update(full_repo)
-                await sourceContext.update(empty_repo)
+                await sourceContext.update(full_record)
+                await sourceContext.update(empty_record)
         async with source as testSource:
             # Open and confirm we saved and loaded correctly
             async with testSource() as sourceContext:
                 with self.subTest(key=full_key):
-                    repo = await sourceContext.repo(full_key)
+                    record = await sourceContext.record(full_key)
                     self.assertEqual(
-                        repo.data.prediction["target_name"]["value"],
+                        record.data.prediction["target_name"]["value"],
                         "feedface",
                     )
                     self.assertEqual(
-                        repo.data.prediction["target_name"]["confidence"], 0.42
+                        record.data.prediction["target_name"]["confidence"],
+                        0.42,
                     )
                 with self.subTest(key=empty_key):
-                    repo = await sourceContext.repo(empty_key)
+                    record = await sourceContext.record(empty_key)
                     self.assertEqual(
                         [
                             val["value"]
-                            for _, val in repo.data.prediction.items()
+                            for _, val in record.data.prediction.items()
                         ],
-                        ["undetermined"] * (len(repo.data.prediction)),
+                        ["undetermined"] * (len(record.data.prediction)),
                     )
                 with self.subTest(both=[full_key, empty_key]):
-                    repos = {
-                        repo.key: repo async for repo in sourceContext.repos()
+                    records = {
+                        record.key: record
+                        async for record in sourceContext.records()
                     }
-                    self.assertIn(full_key, repos)
-                    self.assertIn(empty_key, repos)
+                    self.assertIn(full_key, records)
+                    self.assertIn(empty_key, records)
 
                     self.assertEqual(
-                        repos[full_key].features(), full_repo.features()
+                        records[full_key].features(), full_record.features()
                     )
                     self.assertEqual(
-                        repos[empty_key].features(), empty_repo.features()
+                        records[empty_key].features(), empty_record.features()
                     )
 
 
@@ -134,14 +136,14 @@ class FileSourceTest(SourceTest):
             async with untagged, tagged:
                 async with untagged() as uctx, tagged() as lctx:
                     await uctx.update(
-                        Repo("0", data={"features": {"feed": 1}})
+                        Record("0", data={"features": {"feed": 1}})
                     )
                     await lctx.update(
-                        Repo("0", data={"features": {"face": 2}})
+                        Record("0", data={"features": {"face": 2}})
                     )
             async with untagged, tagged:
                 async with untagged() as uctx, tagged() as lctx:
-                    repo = await uctx.repo("0")
-                    self.assertIn("feed", repo.features())
-                    repo = await lctx.repo("0")
-                    self.assertIn("face", repo.features())
+                    record = await uctx.record("0")
+                    self.assertIn("feed", record.features())
+                    record = await lctx.record("0")
+                    self.assertIn("face", record.features())