+++ examples/source/custom_sqlite.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 from typing import AsyncIterator, NamedTuple, Dict
 
 from dffml.base import BaseConfig
-from dffml.repo import Repo
+from dffml.record import Record
 from dffml.source.source import BaseSourceContext, BaseSource
 from dffml.util.cli.arg import Arg
 
@@ -13,22 +13,22 @@ class CustomSQLiteSourceConfig(BaseConfig, NamedTuple):
 
 
 class CustomSQLiteSourceContext(BaseSourceContext):
-    async def update(self, repo: Repo):
+    async def update(self, record: Record):
         db = self.parent.db
         # Store feature data
         feature_cols = self.parent.FEATURE_COLS
         feature_data = OrderedDict.fromkeys(feature_cols)
-        feature_data.update(repo.features(feature_cols))
+        feature_data.update(record.features(feature_cols))
         await db.execute(
             "INSERT OR REPLACE INTO features (key, "
             + ", ".join(feature_cols)
             + ") "
             "VALUES(?, " + ", ".join("?" * len(feature_cols)) + ")",
-            [repo.key] + list(feature_data.values()),
+            [record.key] + list(feature_data.values()),
         )
         # Store prediction
         try:
-            prediction = repo.prediction("target_name")
+            prediction = record.prediction("target_name")
             prediction_cols = self.parent.PREDICTION_COLS
             prediction_data = OrderedDict.fromkeys(prediction_cols)
             prediction_data.update(prediction.dict())
@@ -37,39 +37,39 @@ class CustomSQLiteSourceContext(BaseSourceContext):
                 + ", ".join(prediction_cols)
                 + ") "
                 "VALUES(?, " + ", ".join("?" * len(prediction_cols)) + ")",
-                [repo.key] + list(prediction_data.values()),
+                [record.key] + list(prediction_data.values()),
             )
         except KeyError:
             pass
 
-    async def repos(self) -> AsyncIterator[Repo]:
+    async def records(self) -> AsyncIterator[Record]:
         # NOTE This logic probably isn't what you want. Only for demo purposes.
         keys = await self.parent.db.execute("SELECT key FROM features")
         for row in await keys.fetchall():
-            yield await self.repo(row["key"])
+            yield await self.record(row["key"])
 
-    async def repo(self, key: str):
+    async def record(self, key: str):
         db = self.parent.db
-        repo = Repo(key)
+        record = Record(key)
         # Get features
         features = await db.execute(
             "SELECT " + ", ".join(self.parent.FEATURE_COLS) + " "
             "FROM features WHERE key=?",
-            (repo.key,),
+            (record.key,),
         )
         features = await features.fetchone()
         if features is not None:
-            repo.evaluated(features)
+            record.evaluated(features)
         # Get prediction
         prediction = await db.execute(
-            "SELECT * FROM prediction WHERE " "key=?", (repo.key,)
+            "SELECT * FROM prediction WHERE " "key=?", (record.key,)
         )
         prediction = await prediction.fetchone()
         if prediction is not None:
-            repo.predicted(
+            record.predicted(
                 "target_name", prediction["value"], prediction["confidence"]
             )
-        return repo
+        return record
 
     async def __aexit__(self, exc_type, exc_value, traceback):
         await self.parent.db.commit()